pulled all navbar-related components out of app.js
